Run benchmark with atm challenges.
